postgre対応強化

+++ migrations/postgres/20250729000004_create_label_counter.sql
@@ -1,7 +1,7 @@
 -- Create label counter table for persistent ID tracking (PostgreSQL compatible)
 CREATE TABLE IF NOT EXISTS label_counter (
-    id INTEGER PRIMARY KEY CHECK (id = 1),
-    current_value INTEGER NOT NULL DEFAULT 0
+    id BIGINT PRIMARY KEY CHECK (id = 1),
+    current_value BIGINT NOT NULL DEFAULT 0
 );
 
 -- Initialize with 0 (will generate 0000 as first ID)
